+++ src/core/node.rs
@@ -1,16 +1,22 @@
 pub struct Node {
-    //coordinates or something
+    pub osm_id: i64,
+    pub lat: f64,
+    pub lon: f64,
 }
 
 impl Node {
     #[inline]
-    pub fn new() -> Self {
-        Node {}
+    pub fn new(osm_id: i64, lat: f64, lon:f64) -> Self {
+        Node {
+            osm_id,
+            lat,
+            lon,
+        }
     }
 }
 
 impl Default for Node {
     fn default() -> Self {
-        Self::new()
+        Self::new(0,0.0,0.0)
     }
 }