+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v0.1.2
+
+### Added
+- Use a kdtree to snap nodes
+- Create a result as wkt
+- make a difference between tower nodes and shape nodes, only use tower nodes for routing
+
+## Changed
+- Vehicle permissions changed
+
 ## v0.1.1
 
 ### Added